feat(EFG-74): Colocando a extensio em arquivo separado
